riemann-client: Make the JSON output optional
Since we require a fairly recent json-c, make the JSON output optional,
so the library can still be compiled without it. The library does not
depend on it anyway, just an example tool, so making it a hard
dependency would be counter productive.

Also remove the json-c installation from the Travis control file, as
Precise does not have a new enough version.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>