simple: New, simplified send & query API
In the most common case, a single event is sent, or a query is launched.
To be able to skip most of the boilerplate in these cases, when the
event to send, or the query to launch can be built with a single
statement, two new functions are introduced: riemann_send() and
riemann_query(). These are available after including <riemann/simple.h>,
and they do the magic behind the curtains.

The README has also been updated to use riemann_send().

Signed-off-by: Gergely Nagy <algernon@balabit.hu>