query: Simplify riemann_query_new()
When creating a new query object, setting the string cannot fail: we
either died before that point, or bailed out when no query string needs
to be set, or we will succeed (or, again, die without recovery).
Therefore, we can skip an error branch there.

Added comments to clarify the situation, however.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>