message: Fix the errno on serialisation failure
When message serialisation fails, we never set an errno, it was
inherited from whatever set it last. To be more in line with other
functions, and to be testable, set errno to EPROTO in
riemann_message_from_buffer(), before calling msg__unpack().

If we return non-NULL, errno should be ignored anyway. But if it returns
NULL, then it will be set to the correct value now.

Also added a test case to trigger this behaviour.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>