Makefile.am: Do not include generated files in the tarball
We want to generate the protobuf-c files during the build, and we do not
want to include them pre-built in the tarball. Including them leads to
interesting conflicts when the system they were generated on had an
older protobuf-c compiler than where it is being compiled. To avoid such
mismatches, we simply don't ship the generated files.

However, hacks had to be made to make sure that all object files still
have the correct dependency, in order not to break parallel builds.

Reported-by: Peter Czanik <czanik@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>