Makefile.am: Improve the doxygen target
Instead of always regenerating the documentation when `make doxygen` is
called, create a new rule called docs/html (the output directory of
doxygen, thus, non-PHONY), which depends on the Doxyfile and the public
headers. This will run doxygen, while the doxygen targets remains PHONY,
and depends on this new one.

This way, documentation is only regenerated, if the sources or the
Doxyfile changed.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>