client: Implement TLS support
This is a big one, with lots of little tricks to keep API and ABI
backwards compatible. With the recent change to riemann_client_connect()
and riemann_client_create(), one is able to pass extra options to these
functions in an extensible way. Since all options are varargs, and
unused when the client type is any of the previously supported types, we
keep both API and ABI stable.

However, if the client type is TLS, we use those options to configure up
our TLS stuff.

This fixes #6.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>