event: Add riemann_event_string_attributes_add()
Implement a convenience function for an easier - and shorter - way of
setting or adding attributes. The new function takes a key-value pair,
creates a riemann_attribute_t object, and attaches it to the event. This
way, one does not need to explicitly create the object.

The riemann_event_set() function was updated to handle the new method of
setting attributes too.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>