simple: Add two new convenience functions.
To make it easier to send a query and receive results with one command,
riemann_communicate_query() was introduced. To send a single (in-place
constructed) event, we now have riemann_communicate_event().

The example application in the README was updated to use the new APIs,
and they also got a mention in NEWS.md.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>