Add a set of version related macros
These can be used to detect the version of the library at compile-time,
without relying on pkg-config. Based on the library version, newer
features of the library may be used. So the purpose of these macros is
to make it easier to optionally use newer features, while still being
compilable with older versions of the library.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>