Adjust the symbol versioning magic
Due to the new enum, we want newly compiled things to depend on the newest
library (because we can't version enums), so `riemann_client_new` and
`riemann_client_create` are now bound to the 1.10 version by default.

We want old binaries to work too, though, so the same functions are also
provided at the old versions too.

Also updated the NEWS file, while there.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>