Rearrange symbol versioning and aliases
To combat a shortcoming in old linkers, rearrange how symbols are versioned,
when there are aliases involved. The main function now is always an alias, to
make things clearer to the compiler.

However, when compiling without symbol versioning, we use the original symbol
names, instead of creating an alias for them.

Fixes #18.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>